requirements.txt
numpy==2.1.1
pydicom==3.0.0
nibabel==5.2.1
pynrrd==1.0.0